# docs/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Identa API
  version: 1.0.0
  description: |
    OpenAPI contract for the current Laravel backend integration.
    This contract is authoritative for frontend/backend alignment.
servers:
  - url: http://localhost:8000
    description: Local backend
...
paths:
...
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: laravel_session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    EnvelopeMeta:
      type: object
      properties:
        request_id:
          type: string
          example: req_01JY2SPWJ5CFVQ68C7W6A41Q6N
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: Validation failed.
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
        meta:
          $ref: "#/components/schemas/EnvelopeMeta"
      required: [error]
...
    ResetPasswordRequest:
      type: object
      required: [token, email, password, password_confirmation]
      properties:
        token:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        password_confirmation:
          type: string
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]